+++ openbas-api/src/main/java/io/openbas/rest/mapper/MapperApi.java
@@ -19,6 +19,7 @@ import io.openbas.rest.mapper.form.ImportMapperUpdateInput;
 import io.openbas.rest.scenario.form.InjectsImportTestInput;
 import io.openbas.rest.scenario.response.ImportPostSummary;
 import io.openbas.rest.scenario.response.ImportTestSummary;
+import io.openbas.service.InjectImportService;
 import io.openbas.service.InjectService;
 import io.openbas.service.MapperService;
 import io.openbas.utils.pagination.SearchPaginationInput;
@@ -51,10 +52,9 @@ import org.springframework.web.reactive.function.UnsupportedMediaTypeException;
 public class MapperApi extends RestBehavior {
 
   private final ImportMapperRepository importMapperRepository;
-
   private final MapperService mapperService;
-
   private final InjectService injectService;
+  private final InjectImportService injectImportService;
 
   // 25mb in byte
   private static final int MAXIMUM_FILE_SIZE_ALLOWED = 25 * 1000 * 1000;
@@ -169,7 +169,7 @@ public class MapperApi extends RestBehavior {
             });
     Scenario scenario = new Scenario();
     scenario.setRecurrenceStart(Instant.now());
-    return injectService.importInjectIntoScenarioFromXLS(
+    return injectImportService.importInjectIntoScenarioFromXLS(
         scenario, importMapper, importId, input.getName(), input.getTimezoneOffset(), false);
   }
